Final TP Project Code for Summer Course of CMU 112
Description of Game:

This was made without Pygame, and such is not as smooth as it would be otherwise. This instead uses Tkinter for the animation framework.
You may also need to download the graphics file from: https://bit.ly/3iuGA2X

The game is called Pong4. Loading up the game you are greeted with a custom start screen. You then have the option to enter the "GameMode" or the "HelpMode". 
"HelpMode" - Consists of a few pages using a Book Class that allows the user to use the arrows to to switch pages, and the whole program is encapsulated in the Mode.
"GameMode"- Using Tkinter, the user is put into a four player pong field. You are on the bottom, and to the top, left, and right are AIs. 
(AIs- Each of these AIs uses the x,y coordinates of both the paddle they are controlling and the ball's x,y coordinates to calculate the balls position after the time determined needed to reach the goal. Then determines if it needs to move the paddle.
User- Using the simple keyPressed function in my graphics file, I take in user input and move the paddle. The collisions mechanics are based off a square to square collision. By using the center of the ball and the radius and the paramters of the paddle, the program will determine if a collision occurs on the top or left or right of the paddle within a certain hitbox
Goals- Created using basic collision code, the program checks if the ball either collides with a wall or hits the goals and results in one of the AIs "dying" or the User "dying".
Interface(on the right, need to expand screen)- Displays the status of the AIs and the time passed, relies on different parts of the program sending information back allowing the interface. 

Features:
1. The user can play with AIs.
2. The collision mechanics are all very smooth
3. There is a way to win, albeit it will take a long time